Adding a benchmark
